+++ controllers/Image.js
@@ -1,0 +1,31 @@
+const Clarifai = require('clarifai');
+const app = new Clarifai.App({
+    apiKey: '6adb3b5cbeb54cd3b14d5605dc3ed5d3',
+   });
+
+const handleAPICall = (req, res) => { 
+    app.models
+        .predict(Clarifai.FACE_DETECT_MODEL, req.body.input)
+        .then(data => { 
+            res.json(data);
+        })
+        .catch(err => res.status(400).json('Unable to work with API'))
+  }
+
+
+const handleImageUpdate = (req, res, db) => { 
+    const { id } = req.body;
+    db('users').where('id', '=', id) 
+        .increment('entries', 1)
+        .returning('entries')
+        .then(entries => { 
+            res.json(entries[0]);
+        })
+    .catch(err => res.status(400).json('unable to get entries'))
+    
+}
+
+module.exports = { 
+    handleImageUpdate,
+    handleAPICall
+};
