+++ controllers/Profile.js
@@ -1,0 +1,17 @@
+const handleProfileGet = (req, res, db) => { 
+    const { id } = req.params;
+    db.select('*').from('users').where({ id })
+    .then(user => { 
+        if (user.length) { 
+            res.json(user[0]);
+        } else { 
+            res.status(400).json('Not found')
+        }
+    })
+    .catch(err => res.status(400).json('error finding user'));
+   
+}
+
+module.exports = { 
+    handleProfileGet
+};
