+++ index.js
@@ -12,6 +12,12 @@ const db = knex({
       database : 'smart-brain'
     }
   });
+const register = require('./controllers/register');
+const signin = require('./controllers/SignIn');
+const profile = require('./controllers/Profile');
+const updateImage = require('./controllers/Image');
+const imageUrl = require('./controllers/Image');
+
 
 const app = express();
 
@@ -20,87 +26,22 @@ app.use(cors());
 
 
 app.get('/', (req, res) => { 
-    res.send(database.users)
+    res.send('success')
 })
 
-app.post('/signin', (req, res) => { 
-    db.select('email', 'hash').from('login')
-        .where('email', '=', req.body.email)
-        .then(data => { 
-            const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-            if (isValid) { 
-               return db.select('*').from('users')
-                .where('email', '=', req.body.email)
-                .then(user => { 
-                    res.json(user[0])
-                })
-                .catch(err => res.status(400).json('wrong credentials'))
-            } else {
+app.post('/signin', (req, res) => { signin.handleSignin(req, res, db, bcrypt)})
 
-                res.status(400).json('wrong credentials')
-            }
-        })
-         .catch(err => res.status(400).json('wrong credentials'))
+app.post('/register', (req, res) => {register.handleRegister(req, res, db, bcrypt)})
 
-})
+app.get('/profile/:id', (req, res) => { profile.handleProfileGet(req, res, db)})
 
-app.post('/register', (req, res) => { 
-    const { email, name, password} = req.body;
-    const hash = bcrypt.hashSync(password);
-    db.transaction(trx => { 
-        trx.insert({ 
-            hash: hash,
-            email: email
-        })
-        .into('login')
-        .returning('email')
-        .then(loginEmail => {
-            return trx('users')
-                .returning('*')
-                .insert({
-                    email: loginEmail[0],
-                    name: name,
-                    joined: new Date()
-                }) 
-                .then(user => { 
-                    res.json(user[0]);
-                })
-        })
-        .then(trx.commit)
-        .catch(trx.rollback)
+app.put('/image', (req, res) => { updateImage.handleImageUpdate(req, res, db)})
 
-    })
-        .catch(err => res.status(400).json('unable to register'))
-})
+app.post('/imageurl', (req, res) => { imageUrl.handleAPICall(req, res)})
 
-app.get('/profile/:id', (req, res) => { 
-    const { id } = req.params;
-    db.select('*').from('users').where({ id })
-    .then(user => { 
-        if (user.length) { 
-            res.json(user[0]);
-        } else { 
-            res.status(400).json('Not found')
-        }
-    })
-    .catch(err => res.status(400).json('error finding user'));
-   
-})
-
-app.put('/image', (req, res) => { 
-    const { id } = req.body;
-    db('users').where('id', '=', id) 
-        .increment('entries', 1)
-        .returning('entries')
-        .then(entries => { 
-            res.json(entries[0]);
-        })
-    .catch(err => res.status(400).json('unable to get entries'))
-    
-})
-
-app.listen(3000, ()=> { 
-    console.log('app is running on port 3000')
+const PORT = process.env.PORT
+app.listen(PORT, ()=> { 
+    console.log(`app is running on port ${PORT}`)
 })
